[jak2] fix missing drill crane center part (#2692)
We were forcing the drill crane to draw at lod 0, but we should have
listened to the game and drawn it at lod 1.

+++ game/system/hid/input_bindings.cpp
@@ -205,7 +205,7 @@ void InputBindingGroups::assign_analog_bind(u32 sdl_idx,
     analog_axii[sdl_idx] = {InputBinding((PadData::AnalogIndex)bind_meta.pad_idx,
                                          bind_meta.for_analog_minimum, modifiers)};
     // there already a bind, so swap (as long as it's not the same key)
-    if (!current_binds.empty() && current_binds.front().sdl_idx != sdl_idx) {
+    if (!current_binds.empty() && (u32)current_binds.front().sdl_idx != sdl_idx) {
       analog_axii[current_binds.front().sdl_idx] = existing_binds;
     }
   } else {
@@ -254,7 +254,7 @@ void InputBindingGroups::assign_button_bind(u32 sdl_idx,
       buttons[sdl_idx] = {InputBinding((PadData::ButtonIndex)bind_meta.pad_idx, modifiers)};
     }
     // there already a bind, so swap (as long as it's not the same key)
-    if (!current_binds.empty() && current_binds.front().sdl_idx != sdl_idx) {
+    if (!current_binds.empty() && current_binds.front().sdl_idx != (s32)sdl_idx) {
       if (current_binds.front().analog_button) {
         button_axii[current_binds.front().sdl_idx] = existing_binds;
       } else {
